- prevent the error because of the creation of two authentication objects with the same name: creator.
- Cleanup the code of SecurityRest.kt

- Move most methods used in SecurityRest.kt to RestIndividualSelectorUtils.kt.
